Added certification actions for suspended/revoked users on activate user btn
Additional activate functions for suspended: check certification and deactivate instead of active if expired and revoked: recertify and activate.
Moved manual recertification to accounts.ex
